+++ internal/management/repository/eventsourcing/view/user_membership.go
@@ -19,7 +19,7 @@ func (v *View) UserMembershipsByAggregateID(aggregateID string) ([]*model.UserMe
 	return view.UserMembershipsByAggregateID(v.Db, userMembershipTable, aggregateID)
 }
 
-func (v *View) SearchUserMemberships(request *usr_model.UserMembershipSearchRequest) ([]*model.UserMembershipView, int, error) {
+func (v *View) SearchUserMemberships(request *usr_model.UserMembershipSearchRequest) ([]*model.UserMembershipView, uint64, error) {
 	return view.SearchUserMemberships(v.Db, userMembershipTable, request)
 }
